continued work on refactoring format_text (wip)
