+++ scripts/constants.jsx
@@ -1,0 +1,163 @@
+// Layer names
+var LayerNames = {
+    WHITE: "W",
+    BLUE: "U",
+    BLACK: "B",
+    RED: "R",
+    GREEN: "G",
+    WU: "WU",
+    UB: "UB",
+    BR: "BR",
+    RG: "RG",
+    GW: "GW",
+    WB: "WB",
+    BG: "BG",
+    GU: "GU",
+    UR: "UR",
+    RW: "RW",
+    ARTIFACT: "Artifact",
+    COLOURLESS: "Colourless",
+    LAND: "Land",
+    GOLD: "Gold",
+    VEHICLE: "Vehicle",
+};
+
+// Font names
+const font_name_mplantin = "MPlantin";
+const font_name_mplantin_italic = "MPlantin-Italic";
+const font_name_ndpmtg = "NDPMTG";
+
+// Font spacing
+const modal_indent = 5.7;
+const line_break_lead = 2.4;
+const flavour_text_lead = 4.4;
+
+// Symbol colours
+var rgb_c = new SolidColor();
+rgb_c.rgb.red = 204;
+rgb_c.rgb.green = 194;
+rgb_c.rgb.blue = 193;
+
+var rgb_w = new SolidColor();
+rgb_w.rgb.red = 255;
+rgb_w.rgb.green = 251;
+rgb_w.rgb.blue = 214;
+
+var rgb_u = new SolidColor();
+rgb_u.rgb.red = 170;
+rgb_u.rgb.green = 224;
+rgb_u.rgb.blue = 250;
+
+var rgb_b = new SolidColor();
+rgb_b.rgb.red = 159;
+rgb_b.rgb.green = 146;
+rgb_b.rgb.blue = 143;
+
+var rgb_r = new SolidColor();
+rgb_r.rgb.red = 249;
+rgb_r.rgb.green = 169;
+rgb_r.rgb.blue = 143;
+
+var rgb_g = new SolidColor();
+rgb_g.rgb.red = 154;
+rgb_g.rgb.green = 211;
+rgb_g.rgb.blue = 175;
+
+// NDPMTG font dictionary to translate Scryfall symbols to font character sequences
+const symbols = {
+    "{W/P}": "Qp",
+    "{U/P}": "Qp",
+    "{B/P}": "Qp",
+    "{R/P}": "Qp",
+    "{G/P}": "Qp",
+    "{E}": "e",
+    "{T}": "ot",
+    "{X}": "ox",
+    "{0}": "o0",
+    "{1}": "o1",
+    "{2}": "o2",
+    "{3}": "o3",
+    "{4}": "o4",
+    "{5}": "o5",
+    "{6}": "o6",
+    "{7}": "o7",
+    "{8}": "o8",
+    "{9}": "o9",
+    "{10}": "oA",
+    "{11}": "oB",
+    "{12}": "oC",
+    "{13}": "oD",
+    "{14}": "oE",
+    "{15}": "oF",
+    "{16}": "oG",
+    "{20}": "oK",
+    "{W}": "ow",
+    "{U}": "ou",
+    "{B}": "ob",
+    "{R}": "or",
+    "{G}": "og",
+    "{C}": "oc",
+    "{W/U}": "QqLS",
+    "{U/B}": "QqMT",
+    "{B/R}": "QqNU",
+    "{R/G}": "QqOV",
+    "{G/W}": "QqPR",
+    "{W/B}": "QqLT",
+    "{B/G}": "QqNV",
+    "{G/U}": "QqPS",
+    "{U/R}": "QqMU",
+    "{R/W}": "QqOR",
+    "{2/W}": "QqWR",
+    "{2/U}": "QqWS",
+    "{2/B}": "QqWT",
+    "{2/R}": "QqWU",
+    "{2/G}": "QqWV",
+    "{S}": "omn",
+    "{Q}": "ol",
+    "{CHAOS}": "?"
+};
+
+// Ability words which should be italicised in formatted text
+const ability_words = [
+    "Adamant",
+    "Addendum",
+    "Battalion",
+    "Bloodrush",
+    "Channel",
+    "Chroma",
+    "Cohort",
+    "Constellation",
+    "Converge",
+    "Council's dilemma",
+    "Delirium",
+    "Domain",
+    "Eminence",
+    "Enrage",
+    "Fateful hour",
+    "Ferocious",
+    "Formidable",
+    "Grandeur",
+    "Hellbent",
+    "Heroic",
+    "Imprint",
+    "Inspired",
+    "Join forces",
+    "Kinship",
+    "Landfall",
+    "Lieutenant",
+    "Metalcraft",
+    "Morbid",
+    "Parley",
+    "Radiance",
+    "Raid",
+    "Rally",
+    "Revolt",
+    "Spell mastery",
+    "Strive",
+    "Sweep",
+    "Tempting offer",
+    "Threshold",
+    "Undergrowth",
+    "Will of the council",
+    "Magecraft"
+];
